Time: 113 ms (65.50%), Space: 26.3 MB (47.01%) - LeetHub

+++ 0015-3sum/0015-3sum.cpp
@@ -14,7 +14,7 @@ public:
                     
                while(j<k)
               {
-                        long long sum=nums[i]+nums[j]+nums[k];
+                        int sum=nums[i]+nums[j]+nums[k];
                   if(sum<0)
                     {
                        j++;
